backend/requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
flask_cors
oauthlib
requests
requests-oauthlib
python-dotenv
